Separate traits from tests in token tests
