+++ tests/src/tests/defuse/env.rs
@@ -1,7 +1,10 @@
 #![allow(dead_code)]
 
-use std::{ops::Deref, sync::LazyLock};
-
+use super::{DefuseExt, accounts::AccountManagerExt};
+use crate::{
+    tests::{defuse::tokens::nep141::traits::DefuseFtReceiver, poa::factory::PoAFactoryExt},
+    utils::{Sandbox, ft::FtExt, read_wasm, wnear::WNearExt},
+};
 use anyhow::anyhow;
 use defuse::{
     contract::{
@@ -15,13 +18,7 @@ use defuse_poa_factory::contract::Role as POAFactoryRole;
 use near_sdk::{AccountId, NearToken};
 use near_workspaces::{Account, Contract, operations::Function};
 use serde_json::json;
-
-use crate::{
-    tests::poa::factory::PoAFactoryExt,
-    utils::{Sandbox, ft::FtExt, read_wasm, wnear::WNearExt},
-};
-
-use super::{DefuseExt, accounts::AccountManagerExt, tokens::nep141::DefuseFtReceiver};
+use std::{ops::Deref, sync::LazyLock};
 
 pub static POA_TOKEN_WASM_NO_REGISTRATION: LazyLock<Vec<u8>> =
     LazyLock::new(|| read_wasm("poa-token-no-registration/defuse_poa_token"));
