+++ tests/src/tests/defuse/tokens/nep141/mod.rs
@@ -1,4 +1,7 @@
+pub mod traits;
+
 use crate::tests::defuse::SigningStandard;
+use crate::tests::defuse::tokens::nep141::traits::DefuseFtWithdrawer;
 use crate::{
     tests::{
         defuse::{DefuseSigner, env::Env},
@@ -8,6 +11,7 @@ use crate::{
 };
 use arbitrary::{Arbitrary, Unstructured};
 use defuse::core::token_id::TokenId;
+use defuse::core::token_id::nep141::Nep141TokenId;
 use defuse::{
     contract::Role,
     core::{
@@ -18,16 +22,13 @@ use defuse::{
 };
 use defuse_randomness::Rng;
 use defuse_test_utils::random::{Seed, random_seed, rng};
-use near_sdk::{AccountId, NearToken, json_types::U128};
+use near_sdk::json_types::U128;
 use rstest::rstest;
-use serde_json::json;
 use std::time::Duration;
 
 #[tokio::test]
 #[rstest]
 async fn deposit_withdraw(#[values(false, true)] no_registration: bool) {
-    use defuse::core::token_id::nep141::Nep141TokenId;
-
     let env = Env::builder()
         .no_registration(no_registration)
         .build()
@@ -110,6 +111,8 @@ async fn poa_deposit(#[values(false, true)] no_registration: bool) {
 #[rstest]
 async fn deposit_withdraw_intent(random_seed: Seed, #[values(false, true)] no_registration: bool) {
     use defuse::core::token_id::nep141::Nep141TokenId;
+
+    use crate::tests::defuse::tokens::nep141::traits::DefuseFtReceiver;
 
     let mut rng = rng(random_seed);
 
@@ -212,7 +215,7 @@ async fn deposit_withdraw_intent_refund(
     use arbitrary::{Arbitrary, Unstructured};
     use defuse::core::token_id::nep141::Nep141TokenId;
 
-    use crate::tests::defuse::SigningStandard;
+    use crate::tests::defuse::{SigningStandard, tokens::nep141::traits::DefuseFtReceiver};
 
     let mut rng = rng(random_seed);
 
@@ -293,6 +296,8 @@ async fn deposit_withdraw_intent_refund(
 async fn ft_force_withdraw(#[values(false, true)] no_registration: bool) {
     use defuse::core::token_id::nep141::Nep141TokenId;
 
+    use crate::tests::defuse::tokens::nep141::traits::DefuseFtWithdrawer;
+
     let env = Env::builder()
         .deployer_as_super_admin()
         .no_registration(no_registration)
@@ -362,146 +367,4 @@ async fn ft_force_withdraw(#[values(false, true)] no_registration: bool) {
             .unwrap(),
         1000
     );
-}
-
-pub trait DefuseFtReceiver {
-    async fn defuse_ft_deposit(
-        &self,
-        defuse_id: &AccountId,
-        token_id: &AccountId,
-        amount: u128,
-        msg: impl Into<Option<DepositMessage>>,
-    ) -> anyhow::Result<u128>;
-}
-
-impl DefuseFtReceiver for near_workspaces::Account {
-    async fn defuse_ft_deposit(
-        &self,
-        defuse_id: &AccountId,
-        token_id: &AccountId,
-        amount: u128,
-        msg: impl Into<Option<DepositMessage>>,
-    ) -> anyhow::Result<u128> {
-        self.ft_transfer_call(
-            token_id,
-            defuse_id,
-            amount,
-            None,
-            &msg.into()
-                .as_ref()
-                .map(ToString::to_string)
-                .unwrap_or_default(),
-        )
-        .await
-    }
-}
-
-impl DefuseFtReceiver for near_workspaces::Contract {
-    async fn defuse_ft_deposit(
-        &self,
-        defuse_id: &AccountId,
-        token_id: &AccountId,
-        amount: u128,
-        msg: impl Into<Option<DepositMessage>>,
-    ) -> anyhow::Result<u128> {
-        self.as_account()
-            .defuse_ft_deposit(defuse_id, token_id, amount, msg)
-            .await
-    }
-}
-
-pub trait DefuseFtWithdrawer {
-    async fn defuse_ft_withdraw(
-        &self,
-        defuse_id: &AccountId,
-        token_id: &AccountId,
-        receiver_id: &AccountId,
-        amount: u128,
-    ) -> anyhow::Result<u128>;
-
-    async fn defuse_ft_force_withdraw(
-        &self,
-        defuse_id: &AccountId,
-        owner_id: &AccountId,
-        token_id: &AccountId,
-        receiver_id: &AccountId,
-        amount: u128,
-    ) -> anyhow::Result<u128>;
-}
-
-impl DefuseFtWithdrawer for near_workspaces::Account {
-    async fn defuse_ft_withdraw(
-        &self,
-        defuse_id: &AccountId,
-        token: &AccountId,
-        receiver_id: &AccountId,
-        amount: u128,
-    ) -> anyhow::Result<u128> {
-        self.call(defuse_id, "ft_withdraw")
-            .deposit(NearToken::from_yoctonear(1))
-            .args_json(json!({
-                "token": token,
-                "receiver_id": receiver_id,
-                "amount": U128(amount),
-            }))
-            .max_gas()
-            .transact()
-            .await?
-            .into_result()?
-            .json::<U128>()
-            .map(|v| v.0)
-            .map_err(Into::into)
-    }
-
-    async fn defuse_ft_force_withdraw(
-        &self,
-        defuse_id: &AccountId,
-        owner_id: &AccountId,
-        token: &AccountId,
-        receiver_id: &AccountId,
-        amount: u128,
-    ) -> anyhow::Result<u128> {
-        self.call(defuse_id, "ft_force_withdraw")
-            .deposit(NearToken::from_yoctonear(1))
-            .args_json(json!({
-                "owner_id": owner_id,
-                "token": token,
-                "receiver_id": receiver_id,
-                "amount": U128(amount),
-            }))
-            .max_gas()
-            .transact()
-            .await?
-            .into_result()?
-            .json::<U128>()
-            .map(|v| v.0)
-            .map_err(Into::into)
-    }
-}
-
-impl DefuseFtWithdrawer for near_workspaces::Contract {
-    async fn defuse_ft_withdraw(
-        &self,
-        defuse_id: &AccountId,
-        token_id: &AccountId,
-        receiver_id: &AccountId,
-        amount: u128,
-    ) -> anyhow::Result<u128> {
-        self.as_account()
-            .defuse_ft_withdraw(defuse_id, token_id, receiver_id, amount)
-            .await
-    }
-
-    async fn defuse_ft_force_withdraw(
-        &self,
-        defuse_id: &AccountId,
-        owner_id: &AccountId,
-        token_id: &AccountId,
-        receiver_id: &AccountId,
-        amount: u128,
-    ) -> anyhow::Result<u128> {
-        self.as_account()
-            .defuse_ft_force_withdraw(defuse_id, owner_id, token_id, receiver_id, amount)
-            .await
-    }
 }
