+++ tests/src/tests/defuse/tokens/nep171/mod.rs
